+++ CHANGELOG.md
@@ -8,8 +8,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [1.0.64] - 2020-06-05
 ### Added
-*   Async Support everywhere
+*   Async support everywhere
 
+### Changed 
+*   Renamed DatabaseAccess class to DB
 
 ### Removed 
 *   Remove IDatabaseAccess interface.